Read and use configured domain-browser mappings
This allows one to configure domain-browser mappings via dconf-editor
in /com/properlypurple/braus/domains

Example mapping:
{'company.com': 'firefox-company.desktop', 'cweiske.de': 'firefox-personal.desktop'}

When braus starts up, it checks if the domain-to-be-opened has a default browser and
directly starts that one. You do not have to select the browser yourself anymore.

Part of: https://github.com/properlypurple/braus/issues/5